Update dfoil.py to accept Dstats = 0
In some cases, D-statistics may legitimately equal 0 (with P=1).

Previous behavior:
DFOIL was checking against the minimum count value, changing P=1 in those cases, and then returning an introgression signature of "na" if P=1.

New behavior:
DFOIL checks against the minimum count value, changing P=-1 in those cases, and returning an introgression signature of "na" if P=-1. Scenarios where a D-stat = 0 with P=1 are treated normally and pass a "0" to the introgression signature.